+++ web/source/package.json
@@ -45,6 +45,10 @@
     "@browserify/envify": "^6.0.0",
     "@browserify/uglifyify": "^6.0.0",
     "@joepie91/eslint-config": "^1.1.1",
+    "@types/bluebird": "^3.5.39",
+    "@types/is-valid-domain": "^0.0.2",
+    "@types/papaparse": "^5.3.9",
+    "@types/psl": "^1.1.1",
     "@types/react-dom": "^18.2.8",
     "@typescript-eslint/eslint-plugin": "^6.7.4",
     "@typescript-eslint/parser": "^6.7.4",
@@ -63,7 +67,10 @@
     "postcss-nested": "^6.0.0",
     "source-map-loader": "^4.0.1",
     "ts-loader": "^9.4.4",
+    "ts-node": "^10.9.1",
+    "ts-patch": "^3.0.2",
     "tsify": "^5.0.4",
-    "typescript": "^5.2.2"
+    "typescript": "^5.2.2",
+    "typia": "^5.1.6"
   }
 }
